requirements.txt
flask==1.1.1
flask-restful
flask-cors
flask-marshmallow
python-dotenv